backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0
pillow>=10.3.0

# NLP & AI
transformers==4.37.2
torch>=2.2.0
spacy>=3.7.2
scikit-learn>=1.4.0
numpy>=1.26.3
pandas>=2.2.0
sentence-transformers>=2.3.1

# Security & Encryption
cryptography>=42.0.2
pycryptodome>=3.20.0

# Firebase
firebase-admin>=6.4.0

# Utilities
pydantic>=2.5.3
pydantic-settings>=2.1.0
aiofiles>=23.2.1
httpx>=0.26.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
